+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 15/04/2017 [Dani 2h]
 
 * Add triggers to app-visuals
-* Add covers to app-controls
+* Add covers to app-control
 * Rename audio to audio-playback
+* app-control query audio-playback for current playing clips
 
 ## 14/04/2017 [Dani 8h]